src/requirements.txt
pydantic
numpy
scipy
pyyaml
